requirements.txt
pytest
requests>=1.0.0
python-dateutil
simplejson
